requirements.txt
paho-mqtt>=1.6.1
joblib>=1.3.2
pandas>=2.0.3
matplotlib>=3.7.1
scikit-learn>=1.3.2